cleanup

+++ core/src/test/java/io/basc/framework/test/RandomTest.java
@@ -16,7 +16,7 @@ public class RandomTest {
 		list.add(3);
 
 		for (int i = 0; i < 1000000; i++) {
-			RandomUtils.random(list, (e) -> (long) e, null);
+			RandomUtils.random(list, (e) -> e, null);
 		}
 	}
 }
